✨ feat(mobile/component): replace update content mutation with reopen content functionality
